+++ SharedStops/Util/Patcher.cs
@@ -1,4 +1,5 @@
 ﻿using HarmonyLib;
+using SharedStopEnabler.StopSelection.Patch;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -19,6 +20,21 @@ namespace SharedStopEnabler.Util
             patched = true;
             var harmony = new Harmony(HarmonyId);
             harmony.PatchAll(Assembly.GetExecutingAssembly());
+
+            ModCompat.ScanMods();
+
+            if (ModCompat.foundMods.ContainsKey(1394468624) && ModCompat.foundMods[1394468624].isEnabled)
+            {
+                Log.Info("AdvancedStopSelection found. Applying patch...");
+
+                MethodInfo original = Type.GetType("ImprovedStopSelection.Detour.TransportToolDetour, ImprovedStopSelection").GetMethod("GetStopPosition", BindingFlags.Instance | BindingFlags.NonPublic);
+                Log.Info($"{original}");
+                MethodInfo postfix = typeof(TransportToolPatch_GetStopPosition).GetMethod("Postfix", BindingFlags.Static | BindingFlags.NonPublic);
+
+                Log.Info($"{original}, {postfix}");
+
+                harmony.Patch(original, postfix: new HarmonyMethod(postfix));
+            }
         }
 
         public static void UnpatchAll()